Differentiate core package and dev requirements (#46)
## Summary (What and Why)

Pytest, aioresponses, and pytest-asyncio are not a requirement of the
package, so we should move it to the dev requirements. This will ensure
no dependency conflicts with `pytest` will affect installers

This change adjusts for this failure by specifying the requirements
directly in the pyproject.toml. It specifies an optional set of `dev`
requirements, just `pytest` for now.

## Test Plan

PR tests

## Release Plan

Owned by online store team